Correct the URL to the collaborators github repo
I renamed the repo, but forgot to update this link to it.

Fortunately, github seem to set up an automatic redirect whenever you
rename a repo, because I've noticed some other links that also still
work.

Anyway, this change corrects the URL to reflect the new name of the
repository.