:whale: Update Dockerfile with health check and additional dependencies
This commit adds a new `HEALTHCHECK` instruction to our Dockerfile which uses curl to monitor the health of our app at regular intervals. The commit also introduces `curl` as an additional dependency, which gets installed during the docker build process.

We moved the declaration of `build-base` and `libffi-dev` each to their own line for better readability. Furthermore, we have separated out the installation of new packages from the existing chain of package updates and upgrades to keep things clean and separate.